Looking for issue
Code was working for classmate Matthew, code was not working for me. he
sent me my code back, I copied and pasted both html and css document
into the file linked with GH in hopes to find the error. — looking to
the left at the change in code it was due to three small errors (not
including the spacing differences we each due), which include line
numbers: 59, 88, 196